Extract the channel grid into web/js/ui/channel-table.js
Rendering, row selection, the row operations (insert/remove/move/copy/cut/
paste), the band-plan presets and the invalid-cell highlighting now live in
one module that owns the selection and invalid-cell state privately. Its own
control and clipboard listeners bind from the module rather than the shared
bindEvents.

Repeater imports and the upload preflight reach it through
rowBuilderHooks/insertRowsAtSelectionOrEnd/applyValidationIssues instead of
sharing closure variables. No user-visible behavior change.
